requirements.txt
Flask
Werkzeug
MarkupSafe
python-dotenv
langchain
langchain-groq
langchain-huggingface
sentence-transformers
transformers
torch
torchvision
faiss-cpu
PyMuPDF
numpy
pandas
scikit-learn
Jinja2
